requirements.txt
# RAG pipeline
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.0.0
streamlit>=1.28.0

# for .env
python-dotenv>=1.0.0
